fix rest routes (#22)
* fix rest routes

* phpcbf
